Implemented the navigation menu
Created new system html, css and javascript
Removed legacy code
It's now needed to declare in every page that uses the nevigation menu, the current menu.
Moved the jquery importing postion from the end of the base.html to the beggining, in order to be active for custom javascript code.