Changing the user model from User to Account
As recommended by Django, Boorunaut is not using the User model anymore. Now, a new model called Account was created and it inherits from the UserAdmin model.
It's possible to modify this model according the project needs.
With this change, any database that uses the old system needs to have the related references modified as well as all the first migrations of each existing app (alternatively, it's possible to just delete them and make a fresh start).