Updated bootstrap and changed cosmo secondary color
Updated to a newer version of bootstrap and popper javascript (it was being used the bootstrap beta version) , and is now being used a local version of the cosmo bootstrap, in order to make future custom modifications.

Cosmo secondary color is equal to the dark color. In order to avoid this redundancy, it was changed all secondary colors to the original bootstrap secondary colors. Other color may have the same problem, as the gray-dark is also equal to dark.